Export protocol and refactor tests
