+++ falcon_pachinko/__init__.py
@@ -2,11 +2,12 @@
 
 from __future__ import annotations
 
-from .resource import WebSocketResource, handles_message
+from .resource import WebSocketLike, WebSocketResource, handles_message
 from .websocket import WebSocketConnectionManager, install
 
 __all__ = (
     "WebSocketConnectionManager",
+    "WebSocketLike",
     "WebSocketResource",
     "handles_message",
     "install",
